Fixed sv_botRace causing mangled configstrings
We can't use Info_SetValueForKey because of its implementation.
Our original infostring looks like this:
    n\Boa\r\2\t\1...
Info_SetValueForKey will call Info_RemoveKey and then add the new
key/value pair at the beginning of the infostring. Unfortunately, the infostrings
for clients differ from "normal" infostrings in that each key/value pair
is key\value\ as opposed to \key\value, so a call to Info_SetValueForKey will make
the client's infostring look like this:
        \r\0n\Boa\t\1...
Which is definitely *not* what we want.